requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
future==0.16.0
idna==2.5
line-bot-sdk==1.3.0
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
virtualenv==15.1.0
